# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zones:
    get:
      operationId: listVirtualZones
      tags: [zones]
      summary: List configurable virtual zones
      responses:
        "200":
          description: Virtual zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VirtualZone"
    post:
      operationId: createVirtualZone
      tags: [zones]
      summary: Create a virtual zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VirtualZoneCreateInput"
      responses:
        "201":
          description: Created virtual zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualZone"
  /zones/{id}:
    patch:
      operationId: updateVirtualZone
      tags: [zones]
      summary: Update a virtual zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VirtualZoneUpdateInput"
      responses:
        "200":
          description: Updated virtual zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualZone"
        "404":
          description: Zone not found
  /zones/{id}/stats:
    get:
      operationId: getVirtualZoneStats
      tags: [zones]
      summary: Get demand statistics for a virtual zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Zone statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualZoneStats"
        "404":
          description: Zone not found
components:
  schemas:
...
    VirtualZone:
      type: object
      required: [id, name, description, centerLatitude, centerLongitude, radiusKm, lowDemandPerHour, highDemandPerHour, minSampleSize, active, createdAt, updatedAt]
      properties:
        id: { type: number }
        name: { type: string, minLength: 2 }
        description: { type: ["string", "null"] }
        centerLatitude: { type: number }
        centerLongitude: { type: number }
        radiusKm: { type: number }
        lowDemandPerHour: { type: number }
        highDemandPerHour: { type: number }
        minSampleSize: { type: number }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    VirtualZoneCreateInput:
      type: object
      required: [name, centerLatitude, centerLongitude, radiusKm, lowDemandPerHour, highDemandPerHour, minSampleSize]
      properties:
        name: { type: string, minLength: 2 }
        description: { type: string }
        centerLatitude: { type: number, minimum: -90, maximum: 90 }
        centerLongitude: { type: number, minimum: -180, maximum: 180 }
        radiusKm: { type: number, minimum: 0.1 }
        lowDemandPerHour: { type: number, minimum: 0 }
        highDemandPerHour: { type: number, minimum: 0 }
        minSampleSize: { type: number, minimum: 1 }
        active: { type: boolean }
    VirtualZoneUpdateInput:
      type: object
      properties:
        name: { type: string, minLength: 2 }
        description: { type: ["string", "null"] }
        centerLatitude: { type: number, minimum: -90, maximum: 90 }
        centerLongitude: { type: number, minimum: -180, maximum: 180 }
        radiusKm: { type: number, minimum: 0.1 }
        lowDemandPerHour: { type: number, minimum: 0 }
        highDemandPerHour: { type: number, minimum: 0 }
        minSampleSize: { type: number, minimum: 1 }
        active: { type: boolean }
    VirtualZoneStats:
      type: object
      required: [zoneId, from, to, demands, completedTrips, demandsPerHour, availableDrivers, activeDrivers, averageSearchTimeSec, averageWaitTimeSec, refusals, cancellations, demandLevel, demandLevelLabel]
      properties:
        zoneId: { type: number }
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        demands: { type: number }
        completedTrips: { type: number }
        demandsPerHour: { type: number }
        availableDrivers: { type: number }
        activeDrivers: { type: number }
        averageSearchTimeSec: { type: ["number", "null"] }
        averageWaitTimeSec: { type: ["number", "null"] }
        refusals: { type: number }
        cancellations: { type: number }
        demandLevel:
          type: string
          enum: [low, medium, high, insufficient]
        demandLevelLabel: { type: string }